requirements.txt
fastapi
uvicorn[standard]
playwright
beautifulsoup4
ics
python-dotenv
lxml
psycopg2-binary
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
celery[redis]
redis
cryptography
pydantic[email]

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
